Rename provider list + schedule, fix UI issues.
