+++ README.md
@@ -1,2 +1,6 @@
 # VT-Care-FE
 Frontend repository for the VT-Care project.
+
+1- Goto directory VTCare/Client
+2- Run "npm install" in CMD in this directory.
+3- Then run "npm run serve -- --port 8081".
